+++ tests/verification/generator_test.py
@@ -1,7 +1,10 @@
 import textwrap
 from pathlib import Path
 
-from rflx.generator.optimizer import optimize
+import pytest
+
+from rflx.generator.optimizer import gnatprove_supports_limit_lines, optimize
+from tests import utils
 
 
 def test_optimize(tmp_path: Path) -> None:
@@ -34,3 +37,13 @@ def test_optimize(tmp_path: Path) -> None:
         end Test;
         """,
     )
+
+
+@pytest.mark.skipif(utils.spark_version() != "24", reason="depends on SPARK 24")
+def test_gnatprove_supports_limit_lines_24() -> None:
+    assert not gnatprove_supports_limit_lines()
+
+
+@pytest.mark.skipif(utils.spark_version() != "25", reason="depends on SPARK 25")
+def test_gnatprove_supports_limit_lines_25() -> None:
+    assert gnatprove_supports_limit_lines()